+++ CHANGELOG.md
@@ -11,7 +11,9 @@
 
 ### Fixed
 
-- broken stack in try expression
+- bugfix code generation of try-finally/defer
+    - fix broken stack unwind in try expression
+    - now maintain try block level
 
 ## [0.28.0] - 2022-12-31